Rewrite the timer system
This is a complete rewrite of the timer system. The previous rudimentary
system was duplicated across the two session components and was unable
to handle improvements such as pausing.

The timer now lives inside of context and lives at the base of the app,
so any component can read and write to it via the `useTimer` hook. The
timer runs a constant loop using set interval, set to 300ms so we can
guarantee the time left will be accurate to the second. The constant
loop means that we don't have to handle starting and stopping. If the
timer is stopped then the interval callback just aborts. The design is
also revised to align more with hooks, so there is less logic in the
loop itself and more useEffect hooks that react to certain changes
(e.g. the newly introduced pause functionality).